codes/raspberry/requirements.txt
google-api-python-client==1.6.5
google-auth-httplib2==0.0.3
google-auth==1.4.1
google-cloud-pubsub==0.32.1
cryptography==2.1.4
pyjwt==1.6.0
paho-mqtt==1.3.1
pathlib
pyserial
json
